added apis documentation

+++ backend/src/modules/roles/controllers/deleteRole.ts
@@ -1,36 +1,26 @@
 import { Request, Response } from "express";
 import { AppDataSource } from "../../../config/data-source.js";
-import {
-	roleRepo,
-	userRepo,
-} from "../../../config/repositories.js";
+import { roleRepo, userRepo } from "../../../config/repositories.js";
 
 export const deleteRole = async (req: Request, res: Response) => {
 	const { id } = req.params as { id: string };
 
-	if (!id) {
-		res.status(400).json({ error: "Invalid ID" });
-		return;
-	}
+	if (!id || isNaN(parseInt(id))) throw Error("Invalid Role Id");
+
+	const roleId = parseInt(id);
 
 	const role = await roleRepo.findOneBy({
-		id: parseInt(id),
+		id: roleId,
 	});
 
-	if (!role) {
-		res.status(404).json({ error: "Invalid ID" });
-		return;
-	}
-    
-	await AppDataSource.transaction(async (transactionalEntityManager) => {
+	if (!role) throw Error("Invalid Role Id");
 
+	await AppDataSource.transaction(async (transactionalEntityManager) => {
 		await transactionalEntityManager.update(
 			userRepo.target,
 			{ role: role.id },
 			{ role: null }
 		);
-
-		
 
 		const result = await transactionalEntityManager.delete(roleRepo.target, {
 			id: role.id,
@@ -39,7 +29,7 @@ export const deleteRole = async (req: Request, res: Response) => {
 		if (result.affected && result.affected > 0) {
 			res.status(204).end();
 		} else {
-			res.status(404).json({ error: "Role not found" });
+			throw Error("Role not found");
 		}
 	});
 };
